requirements.txt
pandas
numpy
langchain==0.3.26
flask==3.1.1  #for frontend and backend
sentence-transformers==4.1.0 #open source embedding model from hugging face
pypdf==5.6.1 #load pdf docs
python-dotenv==1.1.0 
langchain-pinecone==0.2.8 
langchain-openai==0.3.24
langchain-community==0.3.26
sentence_transformers
-e.